Cleaned code/configuration class
