+++ src/main/java/cs222/bsu/edu/wikipedia/GetJSONData.java
@@ -1,7 +1,6 @@
 package cs222.bsu.edu.wikipedia;
 
 
-
 import java.io.IOException;
 import java.io.InputStream;
 import java.net.HttpURLConnection;
@@ -14,24 +13,18 @@ public class GetJSONData {
     public String getSiteData(URL url){
         String json = null;
         try{
-            //connects to url
             HttpURLConnection connection = (HttpURLConnection) url.openConnection();
             connection.setRequestMethod("GET");
             connection.connect();
 
-            //gets json data from url and converts to string
-
             InputStream inStream = connection.getInputStream();
             json = streamToString(inStream);
-
         }
         catch (IOException ex) {
             ex.printStackTrace();
         }
         return json;
-
     }
-    //converts json data to string
     private static String streamToString(InputStream inputStream){
         return new Scanner(inputStream, StandardCharsets.UTF_8).useDelimiter("\\Z").next();
     }
