+++ src/main/java/cs222/bsu/edu/wikipedia/Revision.java
@@ -18,7 +18,7 @@ public class Revision {
     public String timestamp;
     public String user;
     public String comment;
-    public String anon;
+//    public String anon;
 
 
     // a
