+++ src/main/java/cs222/bsu/edu/wikipedia/searchFinder.java
@@ -1,39 +1,16 @@
 package cs222.bsu.edu.wikipedia;
 
-import com.google.gson.GsonBuilder;
-import com.jayway.jsonpath.Configuration;
 import com.jayway.jsonpath.JsonPath;
-import com.jayway.jsonpath.Option;
 import com.jayway.jsonpath.TypeRef;
-import com.jayway.jsonpath.spi.json.GsonJsonProvider;
-import com.jayway.jsonpath.spi.json.JsonProvider;
-import com.jayway.jsonpath.spi.mapper.GsonMappingProvider;
-import com.jayway.jsonpath.spi.mapper.MappingProvider;
 
-import java.util.EnumSet;
 import java.util.List;
-import java.util.Set;
+
 
 public class searchFinder {
     public List<search> getSearchFromJson(String json) {
-        //sets configuration to load/parse JSON
-        Configuration.setDefaults(new Configuration.Defaults() {
+        Configuration configuration = new Configuration();
+        configuration.config();
 
-            private final JsonProvider jsonProvider = new GsonJsonProvider(new GsonBuilder().serializeNulls().create());
-            private final MappingProvider mappingProvider = new GsonMappingProvider();
-
-            public JsonProvider jsonProvider() {
-                return jsonProvider;
-            }
-
-            public MappingProvider mappingProvider() {
-                return mappingProvider;
-            }
-
-            public Set<Option> options() {
-                return EnumSet.noneOf(Option.class);
-            }
-        });
         TypeRef<List<search>> typeRef = new TypeRef<>() {};
 
         return JsonPath.parse(json).read("$..search[0]", typeRef);
